# tests/fixtures/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigationTimelineEventType": {
        "description": "Safe observable workflow actions recorded on an investigation timeline.",
        "enum": [
          "investigation_started",
          "provider_work_started",
          "provider_work_completed",
          "provider_work_failed",
          "evidence_persisted",
          "entities_discovered",
          "pivot_enqueued",
          "pivot_executed",
          "pivot_skipped",
          "research_requested",
          "assessment_requested",
          "investigation_stopped"
        ],
        "title": "InvestigationTimelineEventType",
        "type": "string"
      },
...
      "PageResponse_TimelineEventResponse_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineEventResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "PageResponse[TimelineEventResponse]",
        "type": "object"
      },
...
      "TimelineEventResponse": {
        "additionalProperties": false,
        "description": "One safe observable workflow event.",
        "properties": {
          "entity_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Count"
          },
          "entity_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "evidence_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "pivot_depth": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pivot Depth"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "provider_calls_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Calls Used"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          },
          "relationship_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Relationship Ids",
            "type": "array"
          },
          "replans_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replans Used"
          },
          "target_entity_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Entity Id"
          },
          "type": {
            "$ref": "#/components/schemas/InvestigationTimelineEventType"
          }
        },
        "required": [
          "id",
          "type",
          "occurred_at",
          "provider",
          "target_entity_id",
          "evidence_ids",
          "entity_ids",
          "relationship_ids",
          "error_code",
          "pivot_depth",
          "reason_code",
          "provider_calls_used",
          "replans_used",
          "entity_count"
        ],
        "title": "TimelineEventResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieSession": {
        "in": "cookie",
        "name": "ati_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Agentic Threat Investigator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/investigations/{investigation_id}/timeline": {
      "get": {
        "description": "List observable timeline events in chronological order.",
        "operationId": "list_timeline_events",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "event_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvestigationTimelineEventType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "in": "query",
            "name": "occurred_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurred From"
            }
          },
          {
            "in": "query",
            "name": "occurred_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Occurred To"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_TimelineEventResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Access forbidden."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Resource not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List Timeline Events",
        "tags": [
          "timeline"
        ]
      }
    },
...
  }
}